Prevent appending to stack traces if the size is really large.
Sanity check to ensure stack doesn't get too huge if things go
horribly wrong. Typically this isn't a problem due to stack
overflows, but if you end up with something like a loop between
VMs, the call can be in different threads which avoids that
protection. This doesn't solve that problem, but avoid causing
another with a huge stack trace.