Prevent RemoteCallHandler.invoke with dest of local VMID
This is really a stab in the dark as to a problem that MIGHT be
happening under some unknown circumstances. It has been observed at
least once where an apparently "looping" occurred of supposed remote
calls to the same VM. This should kill that dead, if it's real.
Also includes a special unit test to ensure it didn't break accessing
proxies to "real" objects that live in the local VM.