Ability to force proto 2 and auth failure fixes.
Added an option to `IntrepidSetup` to force the usage of protocol version 2 for
outbound (i.e., client) connections due to a protocol negotiation bug in versions
prior to 1.6.4 which prevents 1.6 from properly negotiating with 1.7.x. When the option
is set, ONLY protocol version 2 will be used. Inbound connections are not effected
by the configuration and may use protocol version 3. (See #18)

User authentication or other session initialization problems generated an NPE
during message encoding resulting in a "Session unexpectedly closed" error.

Fixes: #18