Cap buffer size on send and receive
The VBC ring buffer (for sending) and receive buffer both now have a cap
applied. In the case of the ring buffer, the sender will block if the
ring size grows too large and it will wait for acks to be received.
On the receive size, the window will be set to 0 until the queue size
falls below a low water mark.

Some work has been done on the LargeTransfersTest to allow tweaking the
send buffer size as we hit these buffer filling cases when the send size
is small. However, these changes are incomplete.