Adjust to new mina version as ssl client mode is automatically determined now and a new method must be overrode
Also, the Intrepid builder pattern re-uses the the underlying driver and while the reconnect manager is shutdown its internal thread enters a state the prevents it from being re-uses. The solution was to move the reconnect manager construction into the init of the MINAIntrepidDriver so that a new one is created for every init.